requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
dj-database-url==0.5.0
Django==3.0.1
gunicorn==20.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
http3==0.6.7
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
multidict==4.7.4
mysqlclient==1.4.6
python-decouple==3.3
pytz==2019.3
requests==2.22.0
requests-async==0.6.2
rfc3986==1.3.2
simplejson==3.17.0
sqlparse==0.3.0
urllib3==1.25.7
uvicorn==0.11.2
uvloop==0.14.0
websockets==8.1
whitenoise==5.0.1
yarl==1.4.2
